refactor(cli): rename 'codemod' to 'migrate'
'codemod' implies that only source code is being modified, whereas 'migrate'/'migration' can include
changes to metadata, config, etc... as well as source code
