+++ packages/remix-dev/cli/commands.ts
@@ -20,7 +20,7 @@ import { loadEnv } from "../env";
 import { log } from "./logging";
 import { setupRemix, isSetupPlatform, SetupPlatform } from "./setup";
 
-export * as codemod from "./codemod";
+export * as migrate from "./migrate";
 
 export async function create({
   appTemplate,
