+++ packages/remix-dev/cli/run.ts
@@ -21,7 +21,7 @@ ${colors.heading("Usage")}:
   $ remix dev [${colors.arg("projectDir")}]
   $ remix routes [${colors.arg("projectDir")}]
   $ remix setup [${colors.arg("remixPlatform")}]
-  $ remix codemod <${colors.arg("transform")}> [${colors.arg("projectDir")}]
+  $ remix migrate <${colors.arg("transform")}> [${colors.arg("projectDir")}]
 
 ${colors.heading("Options")}:
   --help, -h          Print this help message and exit
@@ -38,9 +38,9 @@ ${colors.heading("Options")}:
   --debug             Attach Node.js inspector
 \`routes\` Options:
   --json              Print the routes as JSON
-\`codemod\` Options:
+\`migrate\` Options:
   --dry               Dry run (no changes are made to files)
-  --force             Bypass Git safety checks and forcibly run codemods
+  --force             Bypass Git safety checks and forcibly run migration
 
 ${colors.heading("Values")}:
   - ${colors.arg("projectDir")}        The Remix project directory
@@ -48,7 +48,7 @@ ${colors.heading("Values")}:
   - ${colors.arg("remixPlatform")}     \`node\` or \`cloudflare\`
   - ${colors.arg(
     "transform"
-  )}         One of the choices from https://github.com/remix-run/remix/tree/main/packages/remix-dev/cli/codemod/transform-options
+  )}         One of the choices from https://github.com/remix-run/remix/tree/main/packages/remix-dev/cli/migrate/transform-options
 
 ${colors.heading("Creating a new project")}:
 
@@ -312,8 +312,8 @@ export async function run(argv: string[] = process.argv.slice(2)) {
     case "setup":
       await commands.setup(input[1]);
       break;
-    case "codemod": {
-      let answers = await commands.codemod.questions({
+    case "migrate": {
+      let answers = await commands.migrate.questions({
         input: {
           projectDir: input[2],
           transform: input[1],
@@ -321,7 +321,7 @@ export async function run(argv: string[] = process.argv.slice(2)) {
         showHelp,
       });
 
-      await commands.codemod.run({ answers, flags });
+      await commands.migrate.run({ answers, flags });
       break;
     }
     case "dev":
